requirements.txt
Django==5.1.1
django_environ==0.4.5
django_filter==24.2
django_phonenumber_field==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
rest_framework_simplejwt==0.0.1
phonenumbers
sentry-sdk==2.1.1
psycopg2-binary==2.9.3
python-jose
